+++ src/Digipolis.Correlation/CorrelationOptions.cs
@@ -2,7 +2,15 @@
 {
     public class CorrelationOptions
     {
-        public string IdHeaderKey { get; set; } = CorrelationHeaders.IdHeaderKey;
-        public string SourceHeaderKey { get; set; } = CorrelationHeaders.SourceHeaderKey;
+        /// <summary>
+        /// Makes the correlationheader required for all routes that do not match property CorrelationHeaderNotRequiredRouteRegex.
+        /// </summary>
+        public bool CorrelationHeaderRequired { get; set; } = false;
+
+        /// <summary>
+        /// Routes that match this regex will never require a correlation header.
+        /// By default the status routes /vx/status and /status will never require correlation.
+        /// </summary>
+        public string CorrelationHeaderNotRequiredRouteRegex { get; set; } = "^(/v./(?i)(status)/|/(?i)(status)/)";
     }
 }