+++ test/Digipolis.Correlation.UnitTests/HttpClientExtensionsTests.cs
@@ -7,6 +7,8 @@ using System.Net.Http;
 using System.Threading.Tasks;
 using Digipolis.Correlation.UnitTests.Utilities;
 using Xunit;
+using System.Text;
+using Newtonsoft.Json.Linq;
 
 namespace Digipolis.Correlation.UnitTests.CorrelationId
 {
@@ -22,33 +24,7 @@ namespace Digipolis.Correlation.UnitTests.CorrelationId
         }
 
         [Fact]
-        public void AddHeadersToClientWithCorrelationContext()
-        {
-            var client = new HttpClient();
-            var options = new CorrelationOptions();
-            var correlationContext = new CorrelationContext(Utilities.Options.Create(options));
-
-            var id = Guid.NewGuid().ToString();
-            var sourceId = Guid.NewGuid().ToString();
-            var sourceName = "testSource";
-            var instanceId = Guid.NewGuid().ToString();
-            var instanceName = "testSource-instanceName";
-            var userId = "userId";
-            var ipAddress = "194.25.76.122";
-
-            correlationContext.TrySetValues(id, sourceId, sourceName, instanceId, instanceName, userId, ipAddress);
-
-            client.SetCorrelationValues(correlationContext);
-
-            Assert.NotNull(client.DefaultRequestHeaders.Single(h => h.Key == options.IdHeaderKey));
-            Assert.Equal(correlationContext.Id, client.DefaultRequestHeaders.Single(h => h.Key == options.IdHeaderKey).Value.Single());
-
-            Assert.NotNull(client.DefaultRequestHeaders.Single(h => h.Key == options.SourceHeaderKey));
-            Assert.Equal(correlationContext.SourceId.ToString(), client.DefaultRequestHeaders.Single(h => h.Key == options.SourceHeaderKey).Value.Single());
-        }
-
-        [Fact]
-        public void AddHeadersToClientWithServiceProvider()
+        public void AddsCorrelationHeaderWithContext()
         {
             var client = new HttpClient();
             var options = new CorrelationOptions();
@@ -66,11 +42,21 @@ namespace Digipolis.Correlation.UnitTests.CorrelationId
 
             client.SetCorrelationValues(CreateServiceProvider(correlationContext, options));
 
-            Assert.NotNull(client.DefaultRequestHeaders.Single(h => h.Key == options.IdHeaderKey));
-            Assert.Equal(correlationContext.Id.ToString(), client.DefaultRequestHeaders.Single(h => h.Key == options.IdHeaderKey).Value.Single());
+            var result = client.DefaultRequestHeaders.Single(h => h.Key == CorrelationHeaders.HeaderKey);
+            Assert.NotNull(result);
+            byte[] data = Convert.FromBase64String(result.Value.Single());
+            string json = Encoding.UTF8.GetString(data);
 
-            Assert.NotNull(client.DefaultRequestHeaders.Single(h => h.Key == options.SourceHeaderKey));
-            Assert.Equal(correlationContext.SourceId.ToString(), client.DefaultRequestHeaders.Single(h => h.Key == options.SourceHeaderKey).Value.Single());
+            dynamic parsedHeader = JObject.Parse(json);
+            string correlationId = (string)parsedHeader.id;
+
+            Assert.Equal(id, (string)parsedHeader.id);
+            Assert.Equal(sourceId, (string)parsedHeader.sourceId);
+            Assert.Equal(sourceName, (string)parsedHeader.sourceName);
+            Assert.Equal(instanceId, (string)parsedHeader.instanceId);
+            Assert.Equal(instanceName, (string)parsedHeader.instanceName);
+            Assert.Equal(userId, (string)parsedHeader.userId);
+            Assert.Equal(ipAddress, (string)parsedHeader.ipAddress);
         }
 
         private IServiceProvider CreateServiceProvider(ICorrelationContext context, CorrelationOptions options = null)