+++ changelog.md
@@ -1,5 +1,11 @@
 # Correlation Toolbox
 
+## 4.0.0
+
+- Added CorrelationMiddleware that uses the correlationheader to fill the correlationcontext
+- Updated the model and headerkeys used for the correlationheader
+- Use of IApplicationContext to create a new correlationContext
+
 ## 3.0.0
 
 - conversion to csproj and MSBuild.